+++ src/App.js
@@ -7,23 +7,28 @@ import About from './pages/About';
 import NotFound from './pages/NotFound';
 import Home from './pages/Home';
 
+import { GithubProvider } from './context/github/GithubContext';
+
 function App() {
   return (
-    <Router>
-      <div className='flex flex-col justify-between h-screen'>
-        <Navbar />
-        <main className='container mx-auto px-3 pb-12'>
-          <Routes>
-            <Route path='/' element={<Home />} />
-            <Route path='/about' element={<About />} />
-            <Route path='/notfound' element={<NotFound />} />
-            <Route path='/*' element={<NotFound />} />
-          </Routes>
-        </main>
+    <GithubProvider>
+      <Router>
+        <div className='flex flex-col justify-between h-screen'>
+          <Navbar />
 
-        <Footer />
-      </div>
-    </Router>
+          <main className='container mx-auto px-3 pb-12'>
+            <Routes>
+              <Route path='/' element={<Home />} />
+              <Route path='/about' element={<About />} />
+              <Route path='/notfound' element={<NotFound />} />
+              <Route path='/*' element={<NotFound />} />
+            </Routes>
+          </main>
+
+          <Footer />
+        </div>
+      </Router>
+    </GithubProvider>
   );
 }
