+++ src/components/users/UserResults.jsx
@@ -1,28 +1,17 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useContext } from 'react';
 
 import Spinner from '../layout/Spinner';
 import UserItem from './UserItem';
 
+import GithubContext from '../../context/github/GithubContext';
+
 function UserResults() {
 
-    const [ users, setUsers ] = useState([]);
-    const [ loading, setLoading ] = useState(true);
-
-    const fetchUsers = async () => {
-        setLoading(true);
-        const res = await fetch(`${process.env.REACT_APP_GITHUB_URL}/users`, {
-            headers: {
-                'Authorization': `token ${process.env.REACT_APP_GITHUB_TOKEN}`
-            }
-        });
-
-        const data = await res.json();
-        setUsers(data);
-        setLoading(false);
-    }
+    const { users, loading, fetchUsers } = useContext(GithubContext);
 
     useEffect(() => {
         fetchUsers();
+        // eslint-disable-next-line
     }, []);
 
     if(loading) {
